+++ Demo/DemoTests/MoyaProviderIntegrationTests.swift
@@ -1,13 +1,68 @@
 import UIKit
-import Moya
 import Quick
 import Nimble
+import ReactiveMoya
 import OHHTTPStubs
+
+private extension String {
+    var URLEscapedString: String {
+        return self.stringByAddingPercentEncodingWithAllowedCharacters(NSCharacterSet.URLHostAllowedCharacterSet())!
+    }
+}
+
+private enum GitHub {
+    case Zen
+    case UserProfile(String)
+}
+
+extension GitHub : MoyaTarget {
+    var baseURL: NSURL { return NSURL(string: "https://api.github.com")! }
+    var path: String {
+        switch self {
+        case .Zen:
+            return "/zen"
+        case .UserProfile(let name):
+            return "/users/\(name.URLEscapedString)"
+        }
+    }
+    var method: Moya.Method {
+        return .GET
+    }
+    var parameters: [String: AnyObject] {
+        return [:]
+    }
+    var sampleData: NSData {
+        switch self {
+        case .Zen:
+            return "Half measures are as bad as nothing at all.".dataUsingEncoding(NSUTF8StringEncoding)!
+        case .UserProfile(let name):
+            return "{\"login\": \"\(name)\", \"id\": 100}".dataUsingEncoding(NSUTF8StringEncoding)!
+        }
+    }
+}
+
+private func url(route: MoyaTarget) -> String {
+    return route.baseURL.URLByAppendingPathComponent(route.path).absoluteString
+}
+
+private let lazyEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
+    return Endpoint<GitHub>(URL: url(target), sampleResponse: .Closure({.Success(200, {target.sampleData})}), method: target.method, parameters: target.parameters)
+}
+
+private let failureEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
+    let errorData = "Houston, we have a problem".dataUsingEncoding(NSUTF8StringEncoding)!
+    return Endpoint<GitHub>(URL: url(target), sampleResponse: .Error(401, NSError(domain: "com.moya.error", code: 0, userInfo: nil), {errorData}), method: target.method, parameters: target.parameters)
+}
+
 
 func beIndenticalToResponse(expectedValue: MoyaResponse) -> MatcherFunc<MoyaResponse> {
     return MatcherFunc { actualExpression, failureMessage in
-        let instance = actualExpression.evaluate()
-        return instance === expectedValue
+        do {
+            let instance = try actualExpression.evaluate()
+            return instance === expectedValue
+        } catch {
+            return false
+        }
     }
 }
 
@@ -16,23 +71,23 @@ class MoyaProviderIntegrationTests: QuickSpec {
         let userMessage = NSString(data: GitHub.UserProfile("ashfurrow").sampleData, encoding: NSUTF8StringEncoding)
         let zenMessage = NSString(data: GitHub.Zen.sampleData, encoding: NSUTF8StringEncoding)
 
-        beforeEach { () -> () in
+        beforeEach {
             OHHTTPStubs.stubRequestsPassingTest({$0.URL!.path == "/zen"}) { _ in
-                return OHHTTPStubsResponse(data: GitHub.Zen.sampleData, statusCode: 200, headers: nil)
+                return OHHTTPStubsResponse(data: GitHub.Zen.sampleData, statusCode: 200, headers: nil).responseTime(0.5)
             }
 
             OHHTTPStubs.stubRequestsPassingTest({$0.URL!.path == "/users/ashfurrow"}) { _ in
-                return OHHTTPStubsResponse(data: GitHub.UserProfile("ashfurrow").sampleData, statusCode: 200, headers: nil)
+                return OHHTTPStubsResponse(data: GitHub.UserProfile("ashfurrow").sampleData, statusCode: 200, headers: nil).responseTime(0.5)
             }
         }
 
-        afterEach { () -> () in
+        afterEach {
             OHHTTPStubs.removeAllStubs()
         }
 
         describe("valid endpoints") {
             describe("with live data") {
-                describe("a provider") { () -> () in
+                describe("a provider") {
                     var provider: MoyaProvider<GitHub>!
                     beforeEach {
                         provider = MoyaProvider<GitHub>()
@@ -66,14 +121,14 @@ class MoyaProviderIntegrationTests: QuickSpec {
                     }
                     
                     it("returns an error when cancelled") {
-                        var receivedError: NSError?
-                        
+                        var receivedError: ErrorType?
+
                         let target: GitHub = .UserProfile("ashfurrow")
                         let token = provider.request(target) { (data, statusCode, response, error) in
                             receivedError = error
                         }
                         token.cancel()
-                        
+
                         expect(receivedError).toEventuallyNot( beNil() )
                     }
                 }
@@ -81,7 +136,7 @@ class MoyaProviderIntegrationTests: QuickSpec {
                 describe("a provider with network activity closures") {
                     it("notifies at the beginning of network requests") {
                         var called = false
-                        var provider = MoyaProvider<GitHub>(networkActivityClosure: { (change) -> () in
+                        let provider = MoyaProvider<GitHub>(networkActivityClosure: { (change) -> () in
                             if change == .Began {
                                 called = true
                             }
@@ -95,7 +150,7 @@ class MoyaProviderIntegrationTests: QuickSpec {
 
                     it("notifies at the end of network requests") {
                         var called = false
-                        var provider = MoyaProvider<GitHub>(networkActivityClosure: { (change) -> () in
+                        let provider = MoyaProvider<GitHub>(networkActivityClosure: { (change) -> () in
                             if change == .Ended {
                                 called = true
                             }
@@ -108,7 +163,7 @@ class MoyaProviderIntegrationTests: QuickSpec {
                     }
                 }
                 
-                describe("a reactive provider") { () -> () in
+                describe("a reactive provider") {
                     var provider: ReactiveCocoaMoyaProvider<GitHub>!
                     beforeEach {
                         provider = ReactiveCocoaMoyaProvider<GitHub>()